# [NUI] Add ValuePolicy for BindableProperty