"ALL" default BindableProperty.PropertyChanged events what NUI using, dont use 'oldValue'.

But due to that usecase, we always call DefaultValueCreator when we try to set value.

To make NUI platform faster, let we ignore that property getter when we only use setter.

Do not call default creator when we set value

Currently, we only apply this feature to View's bindable property. But in future, we need to create all bindable property policy as IgnoreOldValueWhenSetValue.