@hinohie 2024-01-09
`Tizen.NUI.PerformanceTest` Result

|Creation Type|Before PR (ms)|After PR (ms)|Reduce Rate (1 - after / before)|
|-------------|-------------|-------------|-------------|
| 40 COLOR | 2.1881203703703704 | 1.9591740740740742 | 10.463149075183 % |
| 40 IMAGE | 1.593088888888889 | 1.4098685185185187 | 11.500950866474 % |
| 40 TEXT | 6.198631481481482 | 5.448729629629629 | 12.097861505273 % |
| 40 ROUNDED COLOR | 2.890422222222221 | 2.6508333333333334 | 8.289061959422 % |
| 40 BORDER COLOR | 3.6644759259259265 | 3.549964814814814 | 3.124897350286 % |
| 40 ROUNDED BORDER COLOR | 4.30103888888889 | 4.09705925925926 | 4.742566503097 % |
| 40 BLUR COLOR | 2.54167962962963 | 2.070242592592592 | 18.548247841359 % |
| 40 ROUNDED BLUR COLOR | 2.746287037037037 | 2.284348148148148 | 16.820488268673 % |